+++ src/DataSourceProviders/DataSourceProviderLocal.cs
@@ -1,0 +1,6 @@
+﻿namespace SimpleRag.DataSourceProviders;
+
+/// <summary>
+/// A SourceProvider representing files on a local disk
+/// </summary>
+public class DataSourceProviderLocal : IDataSourceProvider;