+++ src/DataSourceProviders/IDataSourceProvider.cs
@@ -1,0 +1,6 @@
+namespace SimpleRag.DataSourceProviders;
+
+/// <summary>
+/// Interface for a sourceProvider
+/// </summary>
+public interface IDataSourceProvider;