+++ src/DataSources/CSharp/Models/CSharpDataSource.cs
@@ -5,7 +5,7 @@ namespace SimpleRag.DataSources.CSharp.Models;
 /// <summary>
 /// Represent a C# Based Datasource
 /// </summary>
-public abstract class CSharpDataSource : DataSource
+public class CSharpDataSource : DataSource
 {
     /// <summary>
     /// Builder of the desired format of the Content to be vectorized or leave null to use the default provided format