+++ src/Extensions/ServiceCollectionExtensions.cs
@@ -47,8 +47,7 @@ public static class ServiceCollectionExtensions
         services.AddScoped<MarkdownChunker>();
         services.AddScoped<MarkdownDataSourceCommand>();
         services.AddScoped<GitHubQuery>();
-        services.AddScoped<FileContentGitHubQuery>();
-        services.AddScoped<FileContentLocalQuery>();
+        services.AddScoped<FileContentQuery>();
         services.AddSingleton(new GitHubConnection(githubPatToken));
         services.AddScoped<Ingestion>();
         services.AddScoped<Search>();