+++ src/FileContent/Models/FileContentSourceLocal.cs
@@ -1,6 +1,0 @@
-namespace SimpleRag.FileContent.Models;
-
-/// <summary>
-/// Represents a local file content source.
-/// </summary>
-public class FileContentSourceLocal : FileContentSource;